Relocating AmqpException
