+++ azure-eventhubs/src/main/java/com/microsoft/azure/eventhubs/impl/AmqpException.java
@@ -2,7 +2,7 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project root for full license information.
  */
-package com.microsoft.azure.eventhubs;
+package com.microsoft.azure.eventhubs.impl;
 
 import org.apache.qpid.proton.amqp.transport.*;
